requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.6
dropbox==11.36.0

PyPDF2==3.0.1
python-docx==1.1.0

pydantic==2.7.1
typing_extensions==4.8.0
